% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_art_sample.R
\docType{data}
\name{art_sample}
\alias{art_sample}
\title{Art Sample}
\format{
A data frame with 100 rows and 6 variables:
\describe{
  \item{artist}{Artist}
  \item{museum}{Museum}
  \item{gender}{Gender}
  \item{ethnicity}{Ethnicity}
  \item{GEO3major}{Global Environment Outlook GEO3 major region names}
  \item{year}{year}
}
}
\source{
<Topaz, C. M., Klingenberg, B., Turek, D., Heggeseth, B., Harris, P. E., Blackwood, J. C., ... & Murphy, K. M. (2019). Diversity of artists in major US museums. PloS one, 14(3), e0212852.>
}
\usage{
art_sample
}
\description{
Diversity of artists in major U.S. museums
"The Dollar-And-Cents Case Against Hollywood's Exclusion of Women"
\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0212852}.
}
\keyword{datasets}
