% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bechdel.R
\docType{data}
\name{bechdel}
\alias{bechdel}
\title{Bechdel Test for over 1500 movies}
\format{
A data frame with 1794 rows and 3 variables:
\describe{
  \item{year}{The release year of the movie}
  \item{title}{The title of the movie}
  \item{binary}{Bechdel test result (PASS, FAIL)}
}
}
\source{
<https://github.com/fivethirtyeight/data/tree/master/bechdel/>
}
\usage{
bechdel
}
\description{
A dataset containing data behind the story
"The Dollar-And-Cents Case Against Hollywood's Exclusion of Women"
\url{https://fivethirtyeight.com/features/the-dollar-and-cents-case-against-hollywoods-exclusion-of-women/}.
}
\keyword{datasets}
