% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_summary.R
\name{classification_summary}
\alias{classification_summary}
\title{Posterior Classification Summaries}
\usage{
classification_summary(model, data, cutoff = 0.5)
}
\arguments{
\item{model}{an rstanreg model object with binary y}

\item{data}{data frame including the variables in the model}

\item{cutoff}{probability cutoff to classify a new case as positive}
}
\value{

}
\description{
Posterior Classification Summaries
}
