% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_football.R
\docType{data}
\name{football}
\alias{football}
\title{Football Brain Measurements}
\format{

}
\source{
Singh R, Meier T, Kuplicki R, Savitz J, et al., 
"Relationship of Collegiate Football Experience and Concussion 
With Hippocampal Volume and Cognitive Outcome," JAMA, 311(18), 2014
}
\usage{
football
}
\description{
Brain measurements for football and non-football players as provided in the Lock5 package
}
\keyword{datasets}
