% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_classification_summary.R
\name{naive_classification_summary}
\alias{naive_classification_summary}
\title{Posterior Classification Summaries}
\usage{
naive_classification_summary(model, data, y)
}
\arguments{
\item{model}{a naiveBayes model object with categorical y}

\item{data}{data frame including the variables in the model}

\item{y}{a character string indicating the y variable in data}
}
\value{

}
\description{
Posterior Classification Summaries
}
