% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_classification_summary_cv.R
\name{naive_classification_summary_cv}
\alias{naive_classification_summary_cv}
\title{Posterior Classification Summaries}
\usage{
naive_classification_summary_cv(model, data, y, k = 10)
}
\arguments{
\item{model}{a naiveBayes model object with categorical y}

\item{data}{data frame including the variables in the model}

\item{y}{a character string indicating the y variable in data}

\item{k}{the number of folds to use for cross validation}
}
\value{

}
\description{
Posterior Classification Summaries
}
