% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_binomial.R
\name{plot_beta_binomial}
\alias{plot_beta_binomial}
\title{Plot Beta-Binomial Models}
\usage{
plot_beta_binomial(alpha, beta, x = NULL, n = NULL, prior = TRUE,
  likelihood = TRUE, posterior = TRUE)
}
\arguments{
\item{alpha, beta}{shape parameters of the prior Beta distribution.}

\item{x}{number of successes}

\item{n}{number of trials}

\item{prior}{a logical value indicating whether the prior distribution should be plotted.}

\item{likelihood}{a logical value indicating whether the scaled likelihood should be plotted.}

\item{posterior}{a logical value indicating whether posterior distribution should be plotted.}
}
\value{
a ggplot
}
\description{
Plot Beta-Binomial Models
}
\examples{

\dontrun{
plot_beta_binomial(alpha = 1, beta = 13, x = 25, n = 50)
}
\dontrun{
plot_beta_binomial(alpha = 1, beta = 13, x = 25, n = 50, posterior = FALSE)
}
}
