% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_ci.R
\name{plot_beta_ci}
\alias{plot_beta_ci}
\title{Plot Beta Distribution with Credible Interval}
\usage{
plot_beta_ci(alpha, beta, ci_level = 0.95)
}
\arguments{
\item{alpha, beta}{non-negative parameters of the Beta distribution}

\item{ci_level}{credible interval level}
}
\value{
A density plot for the Beta distribution.
}
\description{
Plots the Probability Density Function for Beta Distribution with Credible Interval
}
\examples{
\dontrun{
plot_beta(1,12)
}
}
