% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binomial_likelihood.R
\name{plot_binomial_likelihood}
\alias{plot_binomial_likelihood}
\title{Plot Binomial Likelihood}
\usage{
plot_binomial_likelihood(x, n, mle = FALSE)
}
\arguments{
\item{x}{number of successes}

\item{n}{number of trials}

\item{mle}{a logical value indicating whether maximum likelihood estimate should be plotted}
}
\value{

}
\description{
Plot Binomial Likelihood
}
\examples{
plot_binomial_likelihood(x = 3, n = 10, mle = TRUE)
}
