% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamma.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Plot Gamma Prior/Posterior}
\usage{
plot_gamma(shape, rate, mean = FALSE, mode = FALSE)
}
\arguments{
\item{shape}{A non-negative shape parameter of the Gamma model.}

\item{rate}{A non-negative rate parameter of the Gamma model.}

\item{mean, mode}{take logical values indicating whether mean and the mode of the model should be displayed.}

\item{lambda}{A range of non-negative values for the Gamma model.}
}
\value{
A density plot for the Gamma model.
}
\description{
Plot Gamma Prior/Posterior
}
\examples{
\dontrun{
plot_gamma(shape = 2, rate = 11)
}

}
