% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamma_poisson.R
\name{plot_gamma_poisson}
\alias{plot_gamma_poisson}
\title{Title}
\usage{
plot_gamma_poisson(
  shape,
  rate,
  sum_y = NULL,
  n = NULL,
  prior = TRUE,
  likelihood = TRUE,
  posterior = TRUE
)
}
\arguments{
\item{shape}{A non-negative shape parameter of the Gamma distribution.}

\item{rate}{A non-negative rate parameter of the Gamma distribution.}

\item{sum_y}{sum of observed values for the Poisson likelihood.}

\item{n}{number of observations for the Poisson likelihood.}

\item{prior}{a logical value indicating whether the prior distribution should be plotted.}

\item{likelihood}{a logical value indicating whether the scaled likelihood should be plotted.}

\item{posterior}{a logical value indicating whether posterior distribution should be plotted.}
}
\value{
a ggplot2
}
\description{
Title
}
\examples{
plot_gamma_poisson(shape = 100, rate = 20, sum_y = 39, n = 6)
}
