% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{plot_normal}
\alias{plot_normal}
\title{Plot Normal Prior/Posterior}
\usage{
plot_normal(mean, sd)
}
\arguments{
\item{mean}{mean}

\item{sd}{standard deviation}
}
\value{

}
\description{
Plot Normal Prior/Posterior
}
