% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal_normal.R
\name{plot_normal_normal}
\alias{plot_normal_normal}
\title{Plot Normal-Normal Models}
\usage{
plot_normal_normal(
  mean,
  sd,
  sigma = NULL,
  x_bar = NULL,
  n = NULL,
  prior = TRUE,
  likelihood = TRUE,
  posterior = TRUE
)
}
\arguments{
\item{mean}{prior mean}

\item{sd}{prior standard deviation}

\item{sigma}{likelihood standard deviation}

\item{x_bar}{sample mean of the data}

\item{n}{sample size of the data}

\item{prior}{a logical value indicating whether the prior model should be plotted.}

\item{likelihood}{a logical value indicating whether the scaled likelihood should be plotted.}

\item{posterior}{a logical value indicating whether posterior model should be plotted.}
}
\value{
a ggplot
}
\description{
Plot Normal-Normal Models
}
\examples{

plot_normal_normal(mean = 0, sd = 3, sigma= 4, x_bar = 5, n = 3)
plot_beta_binomial(mean = -1, sd = 5, sigma= 4, x_bar = 5, n = 3, posterior = FALSE)
}
