% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_poisson_gamma.R
\name{plot_poisson_gamma}
\alias{plot_poisson_gamma}
\title{Title}
\usage{
plot_poisson_gamma(
  s,
  r,
  sum_x = NULL,
  n = NULL,
  prior = TRUE,
  likelihood = TRUE,
  posterior = TRUE
)
}
\arguments{
\item{s}{}

\item{r}{}

\item{sum_x}{}

\item{n}{}

\item{prior}{}

\item{likelihood}{}

\item{posterior}{}
}
\value{

}
\description{
Title
}
