% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_poisson_likelihood.R
\name{plot_poisson_likelihood}
\alias{plot_poisson_likelihood}
\title{Plot Poisson Likelihood}
\usage{
plot_poisson_likelihood(x, lambda_upper_bound = 10)
}
\arguments{
\item{x}{vector of observed data}

\item{lambda_upper_bound}{upper bound for lambda}
}
\value{
a ggplot of Poisson likelihood
}
\description{
Plot Poisson Likelihood
}
\examples{
plot_poisson_likelihood(x = c(4, 2, 7), lambda_upper_bound = 10)
}
