% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pop_vs_soda.R
\docType{data}
\name{pop_vs_soda}
\alias{pop_vs_soda}
\title{Pop vs Soda vs Coke}
\format{
A data frame with 374250 observations, one per survey respondent, and 4 variables:
\describe{
  \item{state}{the U.S. state in which the respondent resides}
  \item{region}{region in which the state falls (as defined by the U.S. Census)}
  \item{word_for_cola}{how the respondent refers to fizzy cola drinks}
  \item{pop}{whether or not the respondent refers to fizzy cola drinks as "pop"}
  }
}
\source{
The survey responses were obtained at \url{http://popvssoda.com/} which is maintained by Alan McConchie.
}
\usage{
pop_vs_soda
}
\description{
Results of a volunteer survey on how people around the U.S. refer to fizzy cola drinks. The options are "pop", "soda", "coke", or "other".
}
\keyword{datasets}
