% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_summary.R
\name{prediction_summary}
\alias{prediction_summary}
\title{Posterior Predictive Summaries}
\usage{
prediction_summary(y, yrep, prob_inner = 0.5, prob_outer = 0.95)
}
\arguments{
\item{y}{response variable}

\item{yrep}{predicted reps}

\item{prob_inner}{posterior predictive interval probability}

\item{prob_outer}{posterior predictive interval probability}
}
\value{

}
\description{
Posterior Predictive Summaries
}
