% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_summary_cv.R
\name{prediction_summary_cv}
\alias{prediction_summary_cv}
\title{Cross-Validated Posterior Predictive Summaries}
\usage{
prediction_summary_cv(data, model, k = 10, prob_inner = 0.5, prob_outer = 0.95)
}
\arguments{
\item{data}{data frame including the variables in the model, both response y and predictors x}

\item{model}{an rstanreg model object with quantitative y}

\item{k}{the number of folds to use for cross validation}

\item{prob_inner}{posterior predictive interval probability (a value between 0 and 1)}

\item{prob_outer}{posterior predictive interval probability (a value between 0 and 1)}
}
\value{

}
\description{
Given a set of observed data including a quantitative response variable y 
and an rstanreg model of y, 
this function returns 4 cross-validated measures of the model's posterior prediction quality: 
median absolute prediction error (the typical difference between the observed y values and their posterior predictive medians),
scaled mae (the typical number of absolute deviations between the observed y values and their posterior predictive medians),
the proportion of observed y values that fall within their posterior prediction intervals, the probability levels of which are set by the user.
}
