% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mode.R
\name{sample_mode}
\alias{sample_mode}
\title{Find Sample Mode}
\usage{
sample_mode(x)
}
\arguments{
\item{x}{vector of sample data}
}
\value{
sample mode
}
\description{
Find Sample Mode
}
\examples{
sample_mode(c(3,4,4))
}
