% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_beta_binomial.R
\name{sim_beta_binomial}
\alias{sim_beta_binomial}
\title{Simulate Beta Binomial - RStan wrap}
\usage{
sim_beta_binomial(alpha, beta, x, n, chains = 4, iter = 1000)
}
\arguments{
\item{alpha, beta}{shape parameters of the prior Beta distribution.}

\item{x}{number of successes}

\item{n}{number of trials}

\item{chains}{number of chains}

\item{iter}{number of iterations}
}
\value{

}
\description{
Simulate Beta Binomial - RStan wrap
}
