% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_beta.R
\name{summarize_beta}
\alias{summarize_beta}
\title{Summarize Beta Trends}
\usage{
summarize_beta(alpha, beta)
}
\arguments{
\item{alpha, beta}{shape parameters of the  Beta distribution.}
}
\value{
a summary table
}
\description{
Summarizes the expected value, variance, and mode of the Beta distribution
}
\examples{
\dontrun{
summarize_beta_binomial(1, 15)
}
}
