% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_gamma_poisson.R
\name{summarize_gamma_poisson}
\alias{summarize_gamma_poisson}
\title{Summarize the Gamma-Poisson Model}
\usage{
summarize_gamma_poisson(shape, rate, sum_x = NULL, n = NULL)
}
\arguments{
\item{shape}{A non-negative shape parameter of the Gamma distribution.}

\item{rate}{A non-negative rate parameter of the Gamma distribution.}

\item{sum_x}{sum of observed values for the Poisson likelihood.}

\item{n}{number of observations for the Poisson likelihood.}
}
\value{

}
\description{
Summarize the Gamma-Poisson Model
}
\examples{
#' \dontrun{
summarize_gamma_poisson(3, 4, 7, 12)
}
}
