% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_normal_normal.R
\name{summarize_normal_normal}
\alias{summarize_normal_normal}
\title{Summarize Normal-Normal}
\usage{
summarize_normal_normal(mean, sd, sigma = NULL, y_bar = NULL, n = NULL)
}
\arguments{
\item{mean}{prior mean}

\item{sd}{prior standard deviation}

\item{sigma}{likelihood standard deviation}

\item{y_bar}{sample mean of the data}

\item{n}{sample size of the data}
}
\value{

}
\description{
Summarize Normal-Normal
}
