% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_poisson_gamma.R
\name{summarize_poisson_gamma}
\alias{summarize_poisson_gamma}
\title{Summarize the Poisson-Gamma Model}
\usage{
summarize_poisson_gamma(shape, rate, sum_x = NULL, n = NULL)
}
\arguments{
\item{shape}{A non-negative shape parameter of the Gamma distribution.}

\item{rate}{A non-negative rate parameter of the Gamma distribution.}

\item{sum_x}{sum of observed values for the Poisson likelihood.}

\item{n}{number of observations for the Poisson likelihood.}
}
\description{
Summarize the Poisson-Gamma Model
}
\examples{
#' \dontrun{
summaruze_poisson_gamma(3, 4, 7, 12)
}
}
