% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta.R
\name{plot_beta}
\alias{plot_beta}
\title{Plot Beta Prior/Posterior}
\usage{
plot_beta(alpha, beta)
}
\arguments{
\item{alpha, beta}{non-negative parameters of the Beta distribution}
}
\value{
A \code{\link{ggplot2::ggplot}} with x axis set between 0 and 1
}
\description{
Plots the Probability Density Function for Beta Prior/Posterior
}
\examples{
\dontrun{
plot_beta(1,12)
}
}
