% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta.R
\name{plot_beta}
\alias{plot_beta}
\title{Plot Beta Prior/Posterior}
\usage{
plot_beta(alpha, beta, mean = FALSE, mode = FALSE)
}
\arguments{
\item{alpha, beta}{non-negative parameters of the Beta distribution}

\item{mean, mode}{provides choices whether to display the mean and the mode of the distribution}
}
\value{
A density plot for the Beta distribution.
}
\description{
Plots the Probability Density Function for Beta Prior/Posterior
}
\examples{
\dontrun{
plot_beta(1,12)
}
}
