% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta.R
\name{plot_beta}
\alias{plot_beta}
\title{Plot a Beta Model for \eqn{\pi}}
\usage{
plot_beta(alpha, beta, mean = FALSE, mode = FALSE)
}
\arguments{
\item{alpha, beta}{positive shape parameters of the Beta model}

\item{mean, mode}{a logical value indicating whether to display the model mean and mode}
}
\value{
A density plot for the Beta model.
}
\description{
Plots the probability density function (pdf) for
a Beta(alpha, beta) model of variable \eqn{\pi}.
}
\examples{
plot_beta(alpha = 1, beta = 12, mean = TRUE, mode = TRUE)
}
