% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_ci.R
\name{plot_beta_ci}
\alias{plot_beta_ci}
\title{Plot a Beta Model with Credible Interval}
\usage{
plot_beta_ci(alpha, beta, ci_level = 0.95)
}
\arguments{
\item{alpha, beta}{positive shape parameters of the Beta model}

\item{ci_level}{credible interval level}
}
\value{
A density plot for the Beta model
}
\description{
Plots the probability density function (pdf) for a
Beta(alpha, beta) model of variable \eqn{\pi} with markings indicating
a credible interval for \eqn{\pi}.
}
\examples{
plot_beta_ci(alpha = 7, beta = 12, ci_level = 0.80)
}
