% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binomial_likelihood.R
\name{plot_binomial_likelihood}
\alias{plot_binomial_likelihood}
\title{Plot a Binomial Likelihood Function}
\usage{
plot_binomial_likelihood(y, n, mle = FALSE)
}
\arguments{
\item{y}{number of successes}

\item{n}{number of trials}

\item{mle}{a logical value indicating whether maximum likelihood estimate of \eqn{\pi}, y/n, should be plotted}
}
\value{
a ggplot
}
\description{
Plots the Binomial likelihood function for variable \eqn{\pi}
given y observed successes in a series of n Binomial trials.
}
\examples{
plot_binomial_likelihood(y = 3, n = 10, mle = TRUE)
}
