% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamma.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Plot Gamma Prior/Posterior}
\usage{
plot_gamma(alpha, beta, lambda = c(0, 3))
}
\arguments{
\item{alpha}{A non-negative shape parameter of the Gamma distribution.}

\item{beta}{A non-negative rate parameter of the Gamma distribution.}

\item{lambda}{A range of non-negative values for the Gamma distribution.}
}
\value{
A density plot for the Gamma distribution.
}
\description{
Plot Gamma Prior/Posterior
}
\examples{
\dontrun{
plot_gamma(2,11)
}
}
