% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamma.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Plot Gamma Prior/Posterior}
\usage{
plot_gamma(shape, rate, lambda = c(0, 3), mean = FALSE, mode = FALSE)
}
\arguments{
\item{shape}{A non-negative shape parameter of the Gamma distribution.}

\item{rate}{A non-negative rate parameter of the Gamma distribution.}

\item{lambda}{A range of non-negative values for the Gamma distribution.}

\item{mean, mode}{provides choices whether to display the mean and the mode of the distribution}
}
\value{
A density plot for the Gamma distribution.
}
\description{
Plot Gamma Prior/Posterior
}
\examples{
\dontrun{
plot_gamma(shape = 2, rate = 11)
}

}
