% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gamma.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Plot a Gamma Model for \eqn{\lambda}}
\usage{
plot_gamma(shape, rate, mean = FALSE, mode = FALSE)
}
\arguments{
\item{shape}{non-negative shape parameter of the Gamma model}

\item{rate}{non-negative rate parameter of the Gamma model}

\item{mean, mode}{a logical value indicating whether to display the model mean and mode}
}
\value{
A density plot for the Gamma model.
}
\description{
Plots the probability density function (pdf) for
a Gamma(shape, rate) model of variable \eqn{\lambda}.
}
\examples{
plot_gamma(shape = 2, rate = 11, mean = TRUE, mode = TRUE)

}
