% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normal.R
\name{plot_normal}
\alias{plot_normal}
\title{Plot a Normal Model for \eqn{\mu}}
\usage{
plot_normal(mean, sd)
}
\arguments{
\item{mean}{mean parameter of the Normal model}

\item{sd}{standard deviation parameter of the Normal model}
}
\value{
a ggplot
}
\description{
Plots the probability density function (pdf) for a
Normal(mean, sd^2) model of variable \eqn{\mu}.
}
\examples{
plot_normal(mean = 3.5, sd = 0.5)
}
