% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_summary_cv.R
\name{prediction_summary_cv}
\alias{prediction_summary_cv}
\title{Cross-Validated Posterior Predictive Summaries}
\usage{
prediction_summary_cv(data, model, k = 10, prob_inner = 0.5, prob_outer = 0.95)
}
\arguments{
\item{data}{data frame including the variables in the model}

\item{model}{a rstanreg model object with quantitative y}

\item{k}{the number of folds to use for cross validation}

\item{prob_inner}{posterior predictive interval probability}

\item{prob_outer}{posterior predictive interval probability}
}
\value{

}
\description{
Cross-Validated Posterior Predictive Summaries
}
