% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mode.R
\name{sample_mode}
\alias{sample_mode}
\title{Sample Mode}
\usage{
sample_mode(x)
}
\arguments{
\item{x}{vector of sample data}
}
\value{
sample mode
}
\description{
Calculate the sample mode of vector x.
}
\examples{
sample_mode(rbeta(100, 2, 7))
}
