% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_beta.R
\name{summarize_beta}
\alias{summarize_beta}
\title{Summarize a Beta Model for \eqn{\pi}}
\usage{
summarize_beta(alpha, beta)
}
\arguments{
\item{alpha, beta}{positive shape parameters of the Beta model}
}
\value{
a summary table
}
\description{
Summarizes the expected value, variance, and mode of 
a Beta(alpha, beta) model for variable \eqn{\pi}.
}
\examples{
summarize_beta(alpha = 1, beta = 15)
}
