% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_beta_binomial.R
\name{summarize_beta_binomial}
\alias{summarize_beta_binomial}
\title{Summarize Beta Binomial}
\usage{
summarize_beta_binomial(alpha, beta, x = NULL, n = NULL)
}
\arguments{
\item{alpha, beta}{non-negative parameters of the prior Beta distribution.}

\item{x}{number of successes}

\item{n}{number of trials}
}
\value{
a summary table
}
\description{
Summarizes the mean, and mode of the prior and posterior Beta distributions of \eqn{\pi} which represents the probability of success in a Binomial likelihood.
}
\examples{

summarize_beta_binomial(1, 15, x = 25, n = 50)
}
