% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_gamma.R
\name{summarize_gamma}
\alias{summarize_gamma}
\title{Summarize Gamma Trends}
\usage{
summarize_gamma(shape, rate)
}
\arguments{
\item{shape, rate}{shape and rate parameters of the  Gamma distribution.}
}
\value{
a summary table
}
\description{
Summarizes the expected value, variance, and mode of the Gamma distribution
}
\examples{
\dontrun{
summarize_gamma(1, 15)
}
}
