% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_gamma.R
\name{summarize_gamma}
\alias{summarize_gamma}
\title{Summarize a Gamma Model for \eqn{\lambda}}
\usage{
summarize_gamma(shape, rate)
}
\arguments{
\item{shape}{positive shape parameter of the Gamma model}

\item{rate}{positive rate parameter of the Gamma model}
}
\value{
a summary table
}
\description{
Summarizes the expected value, variance, and mode of 
a Gamma(shape, rate) model for variable \eqn{\lambda}.
}
\examples{
summarize_gamma(shape = 1, rate = 15)

}
